+++ src/markdown_to_devto/utils/exceptions.py
@@ -1,0 +1,26 @@
+class HTTPException(Exception):
+    """Basic exception"""
+
+
+class HTTPAuthException(HTTPException):
+    def __init__(self, msg):
+        self.msg = f"Unable to authenticate request, check `DEVTO_API_KEY`."
+        super().__init__(msg)
+
+
+class HTTPBadRequestException(HTTPException):
+    def __init__(self, msg):
+        self.msg = msg
+        super().__init__(msg)
+
+
+class HTTPServerException(HTTPException):
+    def __init__(self, msg):
+        self.msg = msg
+        super().__init__(msg)
+
+
+class HTTPConnectionException(HTTPException):
+    def __init__(self, msg):
+        self.msg = msg
+        super().__init__(msg)